Use MeterEquivalence in AbstractMeter
