+++ implementations/micrometer-registry-atlas/src/main/java/io/micrometer/atlas/SpectatorCounter.java
@@ -18,7 +18,6 @@ package io.micrometer.atlas;
 import io.micrometer.core.instrument.AbstractMeter;
 import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.Meter;
-import io.micrometer.core.instrument.util.MeterEquivalence;
 
 public class SpectatorCounter extends AbstractMeter implements Counter {
     private com.netflix.spectator.api.Counter counter;
@@ -36,16 +35,5 @@ public class SpectatorCounter extends AbstractMeter implements Counter {
     @Override
     public double count() {
         return counter.count();
-    }
-
-    @SuppressWarnings("EqualsWhichDoesntCheckParameterClass")
-    @Override
-    public boolean equals(Object o) {
-        return MeterEquivalence.equals(this, o);
-    }
-
-    @Override
-    public int hashCode() {
-        return MeterEquivalence.hashCode(this);
     }
 }
