+++ implementations/micrometer-registry-atlas/src/main/java/io/micrometer/atlas/SpectatorGauge.java
@@ -17,7 +17,6 @@ package io.micrometer.atlas;
 
 import io.micrometer.core.instrument.AbstractMeter;
 import io.micrometer.core.instrument.Gauge;
-import io.micrometer.core.instrument.util.MeterEquivalence;
 
 public class SpectatorGauge extends AbstractMeter implements Gauge {
     private com.netflix.spectator.api.Gauge gauge;
@@ -30,16 +29,5 @@ public class SpectatorGauge extends AbstractMeter implements Gauge {
     @Override
     public double value() {
         return gauge.value();
-    }
-
-    @SuppressWarnings("EqualsWhichDoesntCheckParameterClass")
-    @Override
-    public boolean equals(Object o) {
-        return MeterEquivalence.equals(this, o);
-    }
-
-    @Override
-    public int hashCode() {
-        return MeterEquivalence.hashCode(this);
     }
 }
