+++ implementations/micrometer-registry-statsd/src/main/java/io/micrometer/statsd/StatsdCounter.java
@@ -17,7 +17,6 @@ package io.micrometer.statsd;
 
 import io.micrometer.core.instrument.AbstractMeter;
 import io.micrometer.core.instrument.Counter;
-import io.micrometer.core.instrument.util.MeterEquivalence;
 import reactor.core.publisher.FluxSink;
 
 import java.util.concurrent.atomic.DoubleAdder;
@@ -48,17 +47,6 @@ public class StatsdCounter extends AbstractMeter implements Counter {
     @Override
     public double count() {
         return count.doubleValue();
-    }
-
-    @SuppressWarnings("EqualsWhichDoesntCheckParameterClass")
-    @Override
-    public boolean equals(Object o) {
-        return MeterEquivalence.equals(this, o);
-    }
-
-    @Override
-    public int hashCode() {
-        return MeterEquivalence.hashCode(this);
     }
 
     void shutdown() {
