+++ micrometer-core/src/main/java/io/micrometer/core/instrument/AbstractMeter.java
@@ -15,6 +15,15 @@
  */
 package io.micrometer.core.instrument;
 
+import io.micrometer.core.instrument.util.MeterEquivalence;
+import io.micrometer.core.lang.Nullable;
+
+/**
+ * Base class for {@link Meter} implementations.
+ *
+ * @author Jon Schneider
+ * @author Johnny Lim
+ */
 public abstract class AbstractMeter implements Meter {
     private final Meter.Id id;
 
@@ -25,5 +34,16 @@ public abstract class AbstractMeter implements Meter {
     @Override
     public Id getId() {
         return id;
+    }
+
+    @SuppressWarnings("EqualsWhichDoesntCheckParameterClass")
+    @Override
+    public boolean equals(@Nullable Object o) {
+        return MeterEquivalence.equals(this, o);
+    }
+
+    @Override
+    public int hashCode() {
+        return MeterEquivalence.hashCode(this);
     }
 }
